Modify the user model to used in mysqldb and arngodb
