+++ src/main/java/com/linkedin/replica/signing/models/User.java
@@ -1,32 +1,57 @@
-package com.linkedin.replica.signUpInOut.models;
+package com.linkedin.replica.signing.models;
 
+
+import com.arangodb.entity.DocumentField;
+import com.arangodb.velocypack.annotations.Expose;
 
 public class User {
+    @DocumentField(DocumentField.Type.KEY)
     private String id;
-    private String email;
+    private String email, firstName, lastName;
+
+    @Expose(serialize = false, deserialize = false)
     private String password;
-
-
-    public String getPassword() {
-        return password;
-    }
 
     public String getId() {
         return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
     }
 
     public String getEmail() {
         return email;
     }
 
-    public User(String id, String email, String password) {
-        this.id = id;
+    public void setEmail(String email) {
         this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
         this.password = password;
     }
 
-    public User(String email, String password) {
-        this.email = email;
-        this.password = password;
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public User() {
     }
 }
